app/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.8.0
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
